S.B. - Implemented mock feature for similar users recomendation
Values are taken from the itemDataset.csv file, there's still a mismatch issue between the database and the recommendations database.
There's also an issue with some users with id higher than 2300. Because of this, users have been limited to id 2000 or lower for the time being in the similar users recommendation service.